requirements.txt
langchain>=0.1.0
langgraph>=0.0.40
langchain-groq>=0.0.1
langchain-core>=0.1.0
apify-client>=1.7.0
google-api-python-client>=2.108.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
linkedin-api>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.5.0
typing-extensions>=4.8.0
youtube-transcript-api>=0.6.1
pillow>=10.1.0
diffusers>=0.24.0
transformers>=4.35.2
torch>=2.1.1
requests>=2.31.0 
fpdf 